+++ src/components/LinesItems/index.jsx
@@ -1,17 +1,17 @@
 import React from "react";
-import { useDispatch } from "react-redux";
 import PropTypes from "prop-types";
+import { useDispatch } from "react-redux";
 import cx from "classnames";
 import Button from "../@shared/Button";
 import { deleteLinesById } from "../../pages/Lines/slice";
 
-const LineItems = ({ id, name, color }) => {
+const LinesItems = ({ id, name, color }) => {
   const dispatch = useDispatch();
 
   const handleDeleteClick = (event) => {
-    const { name: lineId, value } = event.target;
+    const { name: lineName, value: lineId } = event.target;
 
-    if (window.confirm(`${value}를 삭제하시겠습니까?`)) {
+    if (window.confirm(`${lineName}를 삭제하시겠습니까?`)) {
       dispatch(deleteLinesById(lineId));
     }
   };
@@ -30,8 +30,8 @@ const LineItems = ({ id, name, color }) => {
         theme="icon"
         size="auto"
         onClick={handleDeleteClick}
-        name={id}
-        value={name}
+        name={name}
+        value={id}
       >
         🗑
       </Button>
@@ -39,10 +39,10 @@ const LineItems = ({ id, name, color }) => {
   );
 };
 
-LineItems.propTypes = {
-  id: PropTypes.string.isRequired,
+LinesItems.propTypes = {
+  id: PropTypes.number.isRequired,
   name: PropTypes.string.isRequired,
   color: PropTypes.string.isRequired,
 };
 
-export default LineItems;
+export default LinesItems;
